fix read from wrong queue, infinite loop now
